Remove duplicate code, refactoring.
