+++ src/app/entry-and-visualization/genomic-data-providers/mygeneinfo-search.service.ts
