+++ src/app/entry-and-visualization/genomic-data-providers/utilities/json-navigator.service.ts
