+++ src/app/entry-and-visualization/variant-visualization/clinical-trials/clinical-trials.component.ts
