+++ src/app/entry-and-visualization/variant-visualization/drugs/drug-details-modal.component.ts
