+++ src/app/entry-and-visualization/variant-visualization/drugs/drug.ts
