+++ src/app/entry-and-visualization/variant-visualization/drugs/drugs-search.service.ts
