+++ src/app/entry-and-visualization/variant-visualization/gene/gene-information.component.ts
