+++ src/app/entry-and-visualization/variant-visualization/variant-visualization.component.ts
