+++ src/app/entry-and-visualization/variant-visualization/variant/variant-information.component.ts
